pyproject.toml
[project]
name = "dpolaris-ai"
version = "0.1.0"
description = "dPolaris AI - Trading Intelligence System with Claude Integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "apscheduler>=3.10.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.36",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
dpolaris = "cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
